requirements.txt
mistralai
llmwhisperer-client
python-dotenv